Add ability to edit category manga in v3 API Change category ids to be longs

+++ TachiServer/src/main/java/xyz/nulldev/ts/api/v3/models/categories/WCategory.kt
@@ -1,8 +1,8 @@
 package xyz.nulldev.ts.api.v3.models.categories
 
 data class WCategory(
-        val id: Long,
+        val id: Int,
         val manga: List<Long>,
         val name: String,
-        val order: Long
+        val order: Int
 )
