+++ TachiServer/src/main/java/xyz/nulldev/ts/api/v3/models/categories/WMutateCategoryRequest.kt
@@ -2,5 +2,5 @@ package xyz.nulldev.ts.api.v3.models.categories
 
 data class WMutateCategoryRequest(
         val name: String,
-        val order: Long?
+        val order: Int?
 )
